layout: add a navbar
I want to differentiate blog's entries from more general notes. For
this, we create two menu entries, and add a navbar at the top.

The nav bar let us select between the two kind of articles: blogs or
notes. For now we have a single blog entry, and no notes.

The page to list entries is simplified: we use the same layout for all
lists (tags, notes, blogs).

The CSS is updated to support the new nav bar.