build: get the proper version for tagging
We were evaluating the version of the app when the rule was evaluated,
while we need to get the version after we run `flyctl deploy`.

Since we're tagging a release, let's also annotate the tag.

Remove the `--build-arg` argument to `flyct deploy` since it's being
ignored (same with `docker build`).

Closes #1.