layout: proper TOC
There's a need for two TOCs in the layout: one for when the page is on
mobile; one for non mobile.

When we are on mobile, we display the first TOC, before the article.
When we're not on mobile, we hide that TOC and display one after the
document.

We restructure a bit the layout so that's it's a bit more readable too,
and close tags properly.