ref(build): build and deploy with nix
Refactored the build of the docker image to be done with nix: the flake
knows how to build the docker image, using caddy as a HTTP server. It
generates a small image, with the configuration for caddy and the site
generated by hugo (`nix build`).

Deleted the Dockerfile since the creation is done with nix.

Got rid of the deployment script since this is also done via the
flake (`nix run .#deploy`).