feat(content): add my resume
Add my resume to the content, with a dedicated CSS. The resume is in
org-mode format, and I use `pandoc` to convert it to HTML. This is done
when the site is build, and is integrated in the docker image.

It is available at /resume.html, but is not listed on the site, that way
I can give the URL to who ever is interested in the future.

It would be useful to also generate a PDF version of it and store it
with other static content.